+++ sdk/core/azure-core/azure/core/pipeline/transport/_requests_basic.py
@@ -29,14 +29,15 @@ from typing import Iterator, Optional, Any, Union, TypeVar
 import urllib3 # type: ignore
 from urllib3.util.retry import Retry # type: ignore
 from urllib3.exceptions import (
-    DecodeError, ReadTimeoutError, ProtocolError
+    DecodeError as CoreDecodeError, ReadTimeoutError, ProtocolError
 )
 import requests
 
 from azure.core.configuration import ConnectionConfiguration
 from azure.core.exceptions import (
     ServiceRequestError,
-    ServiceResponseError
+    ServiceResponseError,
+    DecodeError
 )
 from . import HttpRequest # pylint: disable=unused-import
 
@@ -58,11 +59,11 @@ def _read_raw_stream(response, chunk_size=1):
             for chunk in response.raw.stream(chunk_size, decode_content=False):
                 yield chunk
         except ProtocolError as e:
-            raise requests.exceptions.ChunkedEncodingError(e)
-        except DecodeError as e:
-            raise requests.exceptions.ContentDecodingError(e)
+            raise ServiceResponseError(e, error=e)
+        except CoreDecodeError as e:
+            raise DecodeError(e, error=e)
         except ReadTimeoutError as e:
-            raise requests.exceptions.ConnectionError(e)
+            raise ServiceRequestError(e, error=e)
     else:
         # Standard file-like object.
         while True:
@@ -154,6 +155,8 @@ class StreamDownloadGenerator(object):
             raise StopIteration()
         except requests.exceptions.StreamConsumedError:
             raise
+        except requests.exceptions.ContentDecodingError as err:
+            raise DecodeError(err, error=err)
         except Exception as err:
             _LOGGER.warning("Unable to stream download: %s", err)
             self.response.internal_response.close()