+++ sdk/core/azure-core/tests/test_streaming.py
@@ -25,6 +25,7 @@
 # --------------------------------------------------------------------------
 import os
 from azure.core import PipelineClient
+from azure.core.exceptions import DecodeError
 
 def test_decompress_plain_no_header():
     # expect plain text
@@ -90,7 +91,6 @@ def test_compress_compressed_no_header():
 
 def test_decompress_plain_header():
     # expect error
-    import requests
     account_name = "coretests"
     account_url = "https://{}.blob.core.windows.net".format(account_name)
     url = "https://{}.blob.core.windows.net/tests/test_with_header.txt".format(account_name)
@@ -102,7 +102,7 @@ def test_decompress_plain_header():
     try:
         content = b"".join(list(data))
         assert False
-    except requests.exceptions.ContentDecodingError:
+    except DecodeError:
         pass
 
 def test_compress_plain_header():